fix: Remove backslash from punctuators and fix strings
In my previous commit, I added the backslash character to the list of
punctuators, but it is not actually a punctuator. I added this because I
was getting an `UNKNOWN` when the character appeared in certain places
in my string constants, but that was just due to not correctly handling
backslashes embedded in string constants.

I have removed the character from the list of punctuators, and am now
handling it correctly in strings.